requirements.txt
numpy==1.21.5
Pillow==9.1.1
pycocotools==2.0.4
torch==1.11.0
torchvision==0.12.0
